// front/src/app/models/user.ts
export class User {
  id: number;
  username: string;
  email: string;
  first_name: string;
  last_name: string;
  phone: string;
  is_activated: number;
  logined: boolean;
  avatar_path: string;
  join_time: Date;
  last_seen_time: Date;
  read_time: Date;
  notification_read_time: Date;
  password: string;
  discussions_count: number;
  comments_count: number;
  role: string;
  created_at: any;
  updated_at: any;
}